+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unrelease]
-### Changed
-- Updated htslib to version from samtools/dev.
+## [v0.2.1]
 ### Fixed
 - Incorrect processing of non-primary alignments in Python API.
+### Added
+- Add Python packages, available on PyPI.
+### Changed
+- Updated htslib to version from samtools/dev.
+
 
 ## [v0.2.0]
 ### Fixed